Updated script file names
