+++ database/seed.js
@@ -1,46 +1,58 @@
 const faker = require('faker');
-const dbReviewModel = require('./index.js');
+const dbListingModel = require('./index.js/index.js');
 
-const FAKE_DATA_SIZE = 10000;
+// Reviews array for each listing
+let reviewsArr = [];
 
-for (let i = 0; i < FAKE_DATA_SIZE; i++) {
-  const review = {
-    accommodationId: Math.floor(((Math.random() * 100))),
+for (let i = 0; i < 5; i++) {
+  reviewsArr.push({
+    review_id: i,
+    reviewerName: faker.name.findName(),
+    reviewerPhoto: `https://airbnb-reviews-users-pictures.s3-us-west-1.amazonaws.com/${Math.ceil(Math.random() * 3000)}.jpg`,
+    reviewerLink: faker.internet.url(),
+    date: faker.date.recent(),
+    reviewerComment: faker.lorem.paragraphs(),
     scores: {
-      accuracy: Math.round(((Math.random() * 5) + Number.EPSILON + 0.1) * 10) / 10,
-      communication: Math.round(((Math.random() * 5) + Number.EPSILON + 0.1) * 10) / 10,
-      cleanliness: Math.round(((Math.random() * 5) + Number.EPSILON + 0.1) * 10) / 10,
-      checkIn: Math.round(((Math.random() * 5) + Number.EPSILON + 0.1) * 10) / 10,
-      value: Math.round(((Math.random() * 5) + Number.EPSILON + 0.1) * 10) / 10,
-      location: Math.round(((Math.random() * 5) + Number.EPSILON + 0.1) * 10) / 10,
-      outstandingHospitality: Math.random() > 0.5,
-      quickResponses: Math.random() > 0.5,
-      stylishSpace: Math.random() > 0.5,
-      sparklingClean: Math.random() > 0.5,
-      amazingAmenities: Math.random() > 0.5
+      accuracy: faker.finance.amount(1, 5, 1),
+      communication: faker.finance.amount(1, 5, 1),
+      cleanliness: faker.finance.amount(1, 5, 1),
+      checkIn: faker.finance.amount(1, 5, 1),
+      value: faker.finance.amount(1, 5, 1),
+      location: faker.finance.amount(1, 5, 1),
     },
-    reviewAuthorDetails: {
-      name: faker.name.findName(),
-      userPicture: `https://airbnb-reviews-users-pictures.s3-us-west-1.amazonaws.com/${Math.ceil(Math.random() * 3000)}.jpg`,
-      userPageLink: faker.internet.url(),
-      date: faker.date.recent(),
-      reviewText: faker.lorem.paragraphs()
-    }
+  })
+}
+
+// data size 
+const data_size = 10000000;
+// listing and reviews of listing 
+for (let i = 0; i < data_size; i++) {
+  const listings = {
+    listing_id: i,
+    ratingOverall: faker.finance.amount(1,5,1),
+    ratingCommunication: faker.finance.amount(1,5,1),
+    ratingCheckIn: faker.finance.amount(1,5,1),
+    ratingCleanliness: faker.finance.amount(1,5,1),
+    ratingAccuracy: faker.finance.amount(1,5,1),
+    ratingLocation: faker.finance.amount(1,5,1),
+    ratingValue: faker.finance.amount(1,5,1),
+    host: {
+      host_id: i,
+      hostPhoto: `https://airbnb-reviews-users-pictures.s3-us-west-1.amazonaws.com/${Math.ceil(Math.random() * 3000)}.jpg`
+    },
+    reviews: reviewsArr,
   };
-  // fakeReviewsArray.push(review);
-  // console.log(`${i}. `);
-  // console.log(review);
-  // console.log('-----------------------------------------------');
-  // console.log();
+  
+  const listingEntry = new dbListingModel.Listing(listings);
 
-  const reviewEntry = new dbReviewModel.Review(review);
-
-  reviewEntry.save((err, review) => {
-    if (err) {
+  listingEntry.save((err, listing) => {
+    if(err) {
       console.log(err);
     } else {
-      console.log(`${i}.`)
-      console.log(`${reviewEntry._id} has been saved to the db.`);
+      console.log(`Listing #${listingEntry.listing_id} has been saved to the DB!.`);
     }
-  });
-}
+  })
+
+
+};
+
